% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_synergy.R
\name{compute_bridge_recommendations}
\alias{compute_bridge_recommendations}
\title{High-level Bridge Synergy Recommendations}
\usage{
compute_bridge_recommendations(
  cards,
  A_id,
  B_id,
  depth_N = 3L,
  top_k = 50L,
  gamma = 0.65,
  topN = 30L,
  topChains = 5L,
  topPackages = 5L,
  options = list(enforce_color_identity = FALSE, unique_nodes = TRUE, max_candidates =
    2000L, max_chain_len = 6L)
)
}
\arguments{
\item{cards}{List of cards.}

\item{A_id}{Seed A id.}

\item{B_id}{Seed B id.}

\item{depth_N}{Propagation depth.}

\item{top_k}{Neighbor and frontier cap.}

\item{gamma}{Damping factor.}

\item{topN}{Number of bridge cards to keep.}

\item{topChains}{Number of explainable chains to reconstruct.}

\item{topPackages}{Number of packages to build.}

\item{options}{Options list. Defaults:
\code{enforce_color_identity = FALSE}, \code{unique_nodes = TRUE},
\code{max_candidates = 2000L}, \code{max_chain_len = 6L}, \code{support_k = 2L}.}
}
\value{
List with \code{bridges}, \code{chains}, \code{packages}, and \code{debug}.
}
\description{
Compute bridge cards, explainable chains, and package suggestions between two
seed cards.
}
\examples{
cards <- bridge_synergy_example_cards()
rec <- compute_bridge_recommendations(cards, "card_A", "card_B")
stopifnot(is.data.frame(rec$bridges), is.list(rec$chains), is.list(rec$packages))

}
