% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_db.R
\name{db_connect}
\alias{db_connect}
\title{Open a SQLite connection}
\usage{
db_connect(db_path)
}
\arguments{
\item{db_path}{Character scalar. Path to the SQLite database file.}
}
\value{
A \code{DBIConnection} object.
}
\description{
Open a SQLite connection
}
\examples{
\dontrun{
  con <- db_connect("path/to/db.sqlite")
  on.exit(db_disconnect(con), add = TRUE)
}
}
