% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_db.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Safely close a DBI connection}
\usage{
db_disconnect(con)
}
\arguments{
\item{con}{A \code{DBIConnection} object, or \code{NULL}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Safely close a DBI connection
}
\examples{
\dontrun{
  db_disconnect(con)
}
}
