\name{functional_relations}
\alias{functional_relations}
\title{Find Directed Functional Relations}
\usage{functional_relations(source_card, target_card, context = analysis_context())}
\arguments{
\item{source_card}{Source card list.}
\item{target_card}{Target card list.}
\item{context}{Validated analysis context.}
}
\value{A list of directed, evidenced relation records.}
\description{Matches outputs from the source to requirements or costs of the target.}
