% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_synergy.R
\name{build_packages}
\alias{build_packages}
\title{Build Card Packages from Bridge Chains}
\usage{
build_packages(
  bridges,
  chains,
  scoreA,
  scoreB,
  K = 5L,
  support_k = 2L,
  exclude_ids = character()
)
}
\arguments{
\item{bridges}{Bridges data.frame (must contain \code{id}).}

\item{chains}{List of chain objects (\code{bridge_id}, \code{chainA}, \code{chainB}).}

\item{scoreA}{Named numeric scores from seed A propagation.}

\item{scoreB}{Named numeric scores from seed B propagation.}

\item{K}{Number of packages to return.}

\item{support_k}{Number of support cards to append.}

\item{exclude_ids}{Card ids that cannot be added as support cards.}
}
\value{
List of package objects.
}
\description{
Build deck idea packages from bridge chains. For each bridge, package cards
include the union of \code{chainA} and \code{chainB} plus optional support cards ranked
by \code{scoreA + scoreB}.
}
\examples{
cards <- bridge_synergy_example_cards()
rec <- compute_bridge_recommendations(cards, "card_A", "card_B")
pk <- build_packages(rec$bridges, rec$chains, rec$debug$scoreA, rec$debug$scoreB, K = 2L)
stopifnot(is.list(pk))

}
